requirements.txt
absl-py==0.12.0
ansi2html==1.6.0
anyio==2.2.0
appdirs==1.4.4
apt-clone==0.2.1
apturl==0.5.2
arandr==0.1.10
argcomplete==1.10.0
argon2-cffi==20.1.0
astroid==2.5.1
astunparse==1.6.3
async-generator==1.10
attrs==20.3.0
Babel==2.9.0
backcall==0.2.0
beautifulsoup4==4.8.0
bleach==3.3.0
blinker==1.4
blis==0.7.4
boto3==1.17.51
botocore==1.20.51
bpemb==0.3.3
Brlapi==0.7.0
Brotli==1.0.9
cachetools==4.2.1
catalogue==2.0.1
certifi==2019.11.28
cffi==1.14.5
chardet==3.0.4
click==7.1.2
cloudpickle==1.6.0
colorama==0.4.3
command-not-found==0.3
conda==4.3.16
configobj==5.0.6
cryptography==2.8
cupshelpers==1.0
cx-Oracle==8.2.1
cycler==0.10.0
cymem==2.0.5
dash==1.20.0
dash-core-components==1.16.0
dash-html-components==1.1.3
dash-renderer==1.9.1
dash-table==4.11.3
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.0
Deprecated==1.2.12
distlib==0.3.1
distro==1.4.0
distro-info===0.23ubuntu1
docx2txt==0.8
EbookLib==0.17.1
emoji==1.2.0
entrypoints==0.3
es-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/es_core_news_sm-3.0.0/es_core_news_sm-3.0.0-py3-none-any.whl
extract-msg==0.23.1
filelock==3.0.12
findspark==1.4.2
flair==0.8.0.post1
Flask==1.1.2
Flask-Compress==1.9.0
flatbuffers==1.12
ftfy==6.0.1
funcy==1.15
future==0.18.2
gast==0.3.3
gdown==3.12.2
gensim==3.8.3
google-auth==1.28.1
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
graphviz==0.16
grpcio==1.32.0
h5py==2.10.0
httplib2==0.14.0
huggingface-hub==0.0.9
hyperopt==0.2.5
idna==2.8
ifaddr==0.1.6
IMAPClient==2.1.0
IMDbPY==6.8
importlib-metadata==3.10.1
ipykernel==5.5.3
ipython==7.21.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
isort==5.7.0
itsdangerous==1.1.0
Janome==0.4.1
jedi==0.18.0
Jinja2==2.11.3
jmespath==0.10.0
joblib==1.0.1
json5==0.9.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.11
jupyter-console==6.2.0
jupyter-core==4.7.1
jupyter-dash==0.4.0
jupyter-packaging==0.7.12
jupyter-server==1.4.1
jupyterlab==3.0.9
jupyterlab-pygments==0.1.2
jupyterlab-server==2.3.0
jupyterlab-widgets==1.0.0
Keras-Preprocessing==1.1.2
keyring==18.0.1
kiwisolver==1.3.1
konoha==4.6.4
langdetect==1.0.9
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.5.2
linecache2==1.0.0
louis==3.12.0
lxml==4.6.3
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.3.4
MarkupSafe==1.1.0
matplotlib==3.3.4
mccabe==0.6.1
mistune==0.8.4
mpld3==0.3
murmurhash==1.0.5
mysql-connector-python==8.0.25
nbclassic==0.2.6
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nemo-emblems==4.8.1
nest-asyncio==1.5.1
netifaces==0.10.4
networkx==2.5.1
nltk==3.5
nose==1.3.7
notebook==6.2.0
notify2==0.3
numexpr==2.7.3
numpy==1.19.5
oauthlib==3.1.0
olefile==0.46
onboard==1.4.1
opt-einsum==3.3.0
overrides==3.1.0
packaging==20.3
PAM==0.4.2
pandas==1.2.3
pandocfilters==1.4.3
parso==0.8.1
pathy==0.4.0
pbr==5.4.5
pdfminer.six==20181108
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
plotly==4.14.3
preshed==3.0.5
prometheus-client==0.9.0
prompt-toolkit==3.0.16
protobuf==3.15.8
psutil==5.5.1
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycosat==0.6.3
pycparser==2.20
pycrypto==2.6.1
pycryptodome==3.10.1
pycups==1.9.73
pycurl==7.43.0.2
pydantic==1.7.3
Pygments==2.8.0
PyGObject==3.36.0
PyICU==2.4.2
pyinotify==0.9.6
PyJWT==1.7.1
pyLDAvis==3.3.1
pylint==2.7.2
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.6
pyparted==3.11.2
pyRFC3339==1.1
pyrsistent==0.17.3
PySocks==1.7.1
python-apt==2.0.0+ubuntu0.20.4.5
python-dateutil==2.8.1
python-debian===0.1.36ubuntu1
python-Levenshtein==0.12.2
python-pptx==0.6.18
python-xapp==2.0.2
python-xlib==0.23
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
pyzmq==22.0.3
qtconsole==5.0.2
QtPy==1.9.0
regex==2020.11.13
reportlab==3.5.34
requests==2.25.1
requests-file==1.4.3
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
retrying==1.3.3
rsa==4.7.2
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.2
s3transfer==0.3.7
sacremoses==0.0.44
scikit-learn==0.24.1
scipy==1.6.1
seaborn==0.11.1
SecretStorage==2.3.1
segtok==1.5.10
Send2Trash==1.5.0
sentencepiece==0.1.95
setproctitle==1.1.10
simplejson==3.16.0
six==1.12.0
sklearn==0.0
smart-open==3.0.0
sniffio==1.2.0
snscrape==0.3.4
sortedcontainers==2.3.0
soupsieve==1.9.5
spacy==3.0.5
spacy-legacy==3.0.1
SpeechRecognition==3.8.1
sqlitedict==1.7.0
srsly==2.4.0
systemd-python==234
tabulate==0.8.9
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
terminado==0.9.2
testpath==0.4.4
testresources==2.0.0
textblob==0.15.3
textract==1.6.3
thinc==8.0.2
threadpoolctl==2.1.0
tinycss2==1.0.2
tldextract==2.2.1
tokenizers==0.10.2
toml==0.10.2
torch==1.8.1
torchvision==0.9.1
tornado==6.1
tqdm==4.59.0
traceback2==1.4.0
traitlets==5.0.5
transformers==4.5.1
tweepy==3.10.0
typer==0.3.2
typing-extensions==3.7.4.3
tzlocal==1.5.1
ubuntu-advantage-tools==27.0
ubuntu-drivers-common==0.0.0
ufw==0.36
Unidecode==1.1.1
unittest2==1.1.0
urllib3==1.25.8
vaderSentiment==3.3.2
virtualenv==20.4.2
wadllib==1.3.3
wasabi==0.8.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wordcloud==1.8.1
wrapt==1.12.1
xkit==0.0.0
xlrd==1.2.0
XlsxWriter==1.4.3
zeroconf==0.24.4
zipp==3.4.1
